refactor: move number to molecules
